Moved getAbsoluteFilePath to androidshared
