+++ collect_app/src/main/java/org/odk/collect/android/fastexternalitemset/ItemsetDbAdapter.java
@@ -1,5 +1,7 @@
 package org.odk.collect.android.fastexternalitemset;
 
+import static org.odk.collect.androidshared.utils.PathUtils.getAbsoluteFilePath;
+
 import android.content.ContentValues;
 import android.database.Cursor;
 import android.database.SQLException;
@@ -189,7 +191,7 @@ public class ItemsetDbAdapter implements Closeable {
         if (c != null) {
             if (c.getCount() == 1) {
                 c.moveToFirst();
-                String table = getMd5FromString(PathUtils.getAbsoluteFilePath(storagePathProvider.getOdkDirPath(StorageSubdirectory.FORMS), c.getString(c.getColumnIndex(KEY_PATH))));
+                String table = getMd5FromString(getAbsoluteFilePath(storagePathProvider.getOdkDirPath(StorageSubdirectory.FORMS), c.getString(c.getColumnIndex(KEY_PATH))));
                 db.execSQL("DROP TABLE IF EXISTS " + DATABASE_TABLE + table);
             }
             c.close();
